requirements.txt
pyannote.audio>=2.1.1
torch>=1.9.0
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2
python-Levenshtein>=0.12.2
soundfile>=0.10.3
librosa>=0.8.1 